requirements.txt
streamlit
numpy
os
requests
pandas
sqlite3
joblib
thefuzz
plotly.express
time
datetime
sklearn.ensemble
sklearn.model_selection